+++ pom.xml
@@ -4,8 +4,8 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.wss.remediation</groupId>
-    <artifactId>sanitizer</artifactId>
+    <groupId>com.whitesource</groupId>
+    <artifactId>remediation</artifactId>
     <version>1.0.0</version>
 
     <properties>
